mimirpb: make work with new labels.Labels
Create two files with and without the build guard for `-tags stringlabels`.

Stringlabels has a stopgap inefficient implementation for now.
